pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "atb-lib>=0.6.16",
    "azure-identity>=1.25.1",
    "azure-mgmt-costmanagement>=4.0.1",
    "azure-mgmt-resource>=24.0.0",
    "azure-monitor-query>=2.0.0",
    "env-utils>=1.2.0",
    "installer>=0.7.0",
    "numpy>=2.3.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "reportlab>=4.4.4",
    "streamlit>=1.50.0",
]
